+++ src/main/java/capstone/bookitty/BookittyApplication.java
@@ -2,15 +2,16 @@ package capstone.bookitty;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.autoconfigure.domain.EntityScan;
 import org.springframework.cache.annotation.EnableCaching;
 import org.springframework.context.annotation.Bean;
+import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
 import org.springframework.scheduling.annotation.EnableScheduling;
 import org.springframework.web.servlet.config.annotation.CorsRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 
 @SpringBootApplication
-@EnableCaching
-@EnableScheduling
+@EnableCaching @EnableScheduling
 public class BookittyApplication {
 
 	public static void main(String[] args) {
