+++ src/main/java/capstone/bookitty/domain/bookSimilarity/repository/BookSimilarityRepository.java
@@ -1,4 +1,4 @@
-package capstone.bookitty.domain.bookSimilarity.dao;
+package capstone.bookitty.domain.bookSimilarity.repository;
 
 import capstone.bookitty.domain.bookSimilarity.domain.BookSimilarity;
 import capstone.bookitty.domain.bookSimilarity.domain.BookSimilarityId;
@@ -10,7 +10,5 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface BookSimilarityRepository extends JpaRepository<BookSimilarity, BookSimilarityId> {
-    @Query("SELECT bs FROM BookSimilarity bs WHERE bs.isbn1 = :isbn OR bs.isbn2 = :isbn ORDER BY bs.similarity DESC")
-    List<BookSimilarity> findTopSimilarBooks(@Param("isbn") String isbn);
+public interface BookSimilarityRepository extends JpaRepository<BookSimilarity, BookSimilarityId>, BookSimilarityCustomRepository {
 }
