+++ src/main/java/capstone/bookitty/domain/bookState/dto/StateUpdateRequest.java
@@ -1,7 +1,7 @@
 package capstone.bookitty.domain.bookState.dto;
 
 import capstone.bookitty.global.annotation.ValidEnum;
-import capstone.bookitty.domain.bookState.entity.State;
+import capstone.bookitty.domain.bookState.domain.State;
 
 public record StateUpdateRequest(
         @ValidEnum(enumClass = State.class, message = "State is not valid.")
