+++ src/main/java/capstone/bookitty/domain/comment/application/CommentService.java
@@ -11,9 +11,9 @@ import capstone.bookitty.domain.member.domain.Member;
 import capstone.bookitty.domain.comment.exception.CommentNotFoundException;
 import capstone.bookitty.domain.comment.exception.LikeNotFoundException;
 import capstone.bookitty.domain.member.exception.MemberNotFoundException;
-import capstone.bookitty.domain.comment.dao.CommentRepository;
-import capstone.bookitty.domain.comment.dao.LikeRepository;
-import capstone.bookitty.domain.member.dao.MemberRepository;
+import capstone.bookitty.domain.comment.repository.CommentRepository;
+import capstone.bookitty.domain.comment.repository.LikeRepository;
+import capstone.bookitty.domain.member.repository.MemberRepository;
 import jakarta.persistence.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
