+++ src/main/java/capstone/bookitty/domain/comment/repository/CommentRepository.java
@@ -1,4 +1,4 @@
-package capstone.bookitty.domain.comment.dao;
+package capstone.bookitty.domain.comment.repository;
 
 import capstone.bookitty.domain.comment.domain.Comment;
 import org.springframework.data.domain.Page;
