+++ src/main/java/capstone/bookitty/domain/comment/repository/LikeRepository.java
@@ -1,4 +1,4 @@
-package capstone.bookitty.domain.comment.dao;
+package capstone.bookitty.domain.comment.repository;
 
 import capstone.bookitty.domain.comment.domain.Like;
 import org.springframework.data.jpa.repository.JpaRepository;
