+++ src/main/java/capstone/bookitty/domain/star/application/StarService.java
@@ -8,8 +8,8 @@ import capstone.bookitty.domain.member.domain.Member;
 import capstone.bookitty.domain.star.domain.Star;
 import capstone.bookitty.domain.member.exception.MemberNotFoundException;
 import capstone.bookitty.domain.star.exception.StarNotFoundException;
-import capstone.bookitty.domain.member.dao.MemberRepository;
-import capstone.bookitty.domain.star.dao.StarRepository;
+import capstone.bookitty.domain.member.repository.MemberRepository;
+import capstone.bookitty.domain.star.repository.StarRepository;
 import jakarta.persistence.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
