+++ src/test/java/capstone/bookitty/BookittyApplicationTests.java
@@ -1,7 +1,10 @@
 package capstone.bookitty;
 
 import org.junit.jupiter.api.Test;
+import org.springframework.boot.autoconfigure.ImportAutoConfiguration;
+import org.springframework.boot.autoconfigure.batch.BatchAutoConfiguration;
 import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.transaction.annotation.Transactional;
 
 @SpringBootTest
 class BookittyApplicationTests {
