+++ src/test/java/capstone/bookitty/controllerTest/BookStateControllerTest.java
@@ -2,9 +2,9 @@ package capstone.bookitty.controllerTest;
 
 import capstone.bookitty.IntergrationTest;
 import capstone.bookitty.domain.bookState.dto.StateSaveRequest;
-import capstone.bookitty.domain.bookState.entity.BookState;
+import capstone.bookitty.domain.bookState.domain.BookState;
 import capstone.bookitty.domain.member.domain.Member;
-import capstone.bookitty.domain.bookState.entity.State;
+import capstone.bookitty.domain.bookState.domain.State;
 import org.junit.jupiter.api.Test;
 import org.springframework.http.MediaType;
 import org.springframework.test.web.servlet.ResultActions;
